requirements_transcriber.txt
fastapi
gunicorn
uvicorn[standard]
whisper
starlette
torch
python-dotenv
